pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "formalguard"
version = "0.1.0"
description = "Open-source formal verification framework for security-critical financial hardware"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Thuvaragan Paramsothy"},
    {name = "Sanchayan Paramsothy"},
    {name = "Jananthan Paramsothy"},
]
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.0",
    "rich>=13.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
fg-report = "tools.compliance_report:cli"
fg-lint = "tools.property_lint:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
